+++ adapter/amazon/src/main/java/org/bidon/amazon/SlotType.kt
@@ -1,0 +1,24 @@
+package org.bidon.amazon
+
+/**
+ * Created by Aleksei Cherniaev on 27/09/2023.
+ */
+enum class SlotType(val format: String) {
+    /**
+     * All banner sizes
+     */
+    BANNER("BANNER"),
+    MREC("MREC"),
+
+    REWARDED_AD("REWARDED"),
+
+    /**
+     * Interstitial
+     */
+    VIDEO("VIDEO"),
+    INTERSTITIAL("INTERSTITIAL");
+
+    companion object {
+        fun get(format: String): SlotType? = values().firstOrNull { it.format == format }
+    }
+}