+++ adapter/bigoads/src/main/java/org/bidon/bigoads/impl/BigoAdsBannerImpl.kt
@@ -13,6 +13,7 @@ import org.bidon.sdk.adapter.impl.AdEventFlowImpl
 import org.bidon.sdk.ads.banner.BannerFormat
 import org.bidon.sdk.ads.banner.helper.getHeightDp
 import org.bidon.sdk.ads.banner.helper.getWidthDp
+import org.bidon.sdk.auction.AdTypeParam
 import org.bidon.sdk.config.BidonError
 import org.bidon.sdk.logs.analytic.AdValue
 import org.bidon.sdk.logs.analytic.AdValue.Companion.USD
@@ -73,7 +74,7 @@ internal class BigoAdsBannerImpl :
         return AdViewHolder(bannerAd.adView(), width, height)
     }
 
-    override suspend fun getToken(context: Context): String? = BigoAdSdk.getBidderToken()
+    override suspend fun getToken(context: Context, adTypeParam: AdTypeParam): String? = BigoAdSdk.getBidderToken()
 
     override fun load(adParams: BigoBannerAuctionParams) {
         val builder = BannerAdRequest.Builder()
@@ -89,20 +90,17 @@ internal class BigoAdsBannerImpl :
                     BannerFormat.LeaderBoard -> AdSize.BANNER
                 }
             )
-        val loader = BannerAdLoader.Builder().withAdLoadListener(object : AdLoadListener<BannerAd> {
-            override fun onError(adError: AdError) {
-                val error = adError.asBidonError()
-                logError(TAG, "Error while loading ad: $adError. $this", error)
-                emitEvent(AdEvent.LoadFailed(BidonError.NoFill(demandId)))
-            }
+        val loader = BannerAdLoader.Builder()
+            .withAdLoadListener(object : AdLoadListener<BannerAd> {
+                override fun onError(adError: AdError) {
+                    val error = adError.asBidonError()
+                    logError(TAG, "Error while loading ad: $adError. $this", error)
+                    emitEvent(AdEvent.LoadFailed(BidonError.NoFill(demandId)))
+                }
 
-            override fun onAdLoaded(bannerAd: BannerAd) {
-                logInfo(TAG, "onAdLoaded: $bannerAd, $this")
-                this@BigoAdsBannerImpl.bannerAd = bannerAd
-                val ad = getAd(this)
-                if (ad == null) {
-                    emitEvent(AdEvent.ShowFailed(BidonError.AdNotReady))
-                } else {
+                override fun onAdLoaded(bannerAd: BannerAd) {
+                    logInfo(TAG, "onAdLoaded: $bannerAd, $this")
+                    this@BigoAdsBannerImpl.bannerAd = bannerAd
                     bannerAd.setAdInteractionListener(object : AdInteractionListener {
                         override fun onAdError(error: AdError) {
                             val cause = error.asBidonError()
@@ -113,33 +111,42 @@ internal class BigoAdsBannerImpl :
                         override fun onAdImpression() {
                             logInfo(TAG, "onAdImpression: $this")
                             // tracked impression/shown by [BannerView]
-                            emitEvent(
-                                AdEvent.PaidRevenue(
-                                    ad = ad,
-                                    adValue = AdValue(
-                                        adRevenue = adParams.bidPrice / 1000.0,
-                                        precision = Precision.Precise,
-                                        currency = USD,
+                            getAd()?.let { ad ->
+                                emitEvent(
+                                    AdEvent.PaidRevenue(
+                                        ad = ad,
+                                        adValue = AdValue(
+                                            adRevenue = adParams.bidPrice / 1000.0,
+                                            precision = Precision.Precise,
+                                            currency = USD,
+                                        )
                                     )
                                 )
-                            )
+                            }
                         }
 
                         override fun onAdClicked() {
                             logInfo(TAG, "onAdClicked: $this")
-                            emitEvent(AdEvent.Clicked(ad))
+                            getAd()?.let { ad ->
+                                emitEvent(AdEvent.Clicked(ad))
+                            }
                         }
 
                         override fun onAdOpened() {}
                         override fun onAdClosed() {}
                     })
-                    emitEvent(AdEvent.Fill(ad))
+                    getAd()?.let { ad ->
+                        emitEvent(AdEvent.Fill(ad))
+                    }
                 }
-            }
-        })
+            })
         adParams.activity.runOnUiThread {
-            loader.build()
-                .loadAd(builder.build())
+            loader
+                .build()
+                .loadAd(
+                    builder
+                        .build()
+                )
         }
     }
 }