+++ adapter/bigoads/src/main/java/org/bidon/bigoads/impl/BigoAdsInterstitialImpl.kt
@@ -10,6 +10,7 @@ import org.bidon.sdk.adapter.AdSource
 import org.bidon.sdk.adapter.Mode
 import org.bidon.sdk.adapter.impl.AdEventFlow
 import org.bidon.sdk.adapter.impl.AdEventFlowImpl
+import org.bidon.sdk.auction.AdTypeParam
 import org.bidon.sdk.config.BidonError
 import org.bidon.sdk.logs.analytic.AdValue
 import org.bidon.sdk.logs.analytic.Precision
@@ -39,10 +40,7 @@ internal class BigoAdsInterstitialImpl :
     override val isAdReadyToShow: Boolean
         get() = interstitialAd != null && interstitialAd?.isExpired != false
 
-    override fun destroy() {
-        interstitialAd?.destroy()
-        interstitialAd = null
-    }
+    override suspend fun getToken(context: Context, adTypeParam: AdTypeParam): String? = BigoAdSdk.getBidderToken()
 
     override fun getAuctionParam(auctionParamsScope: AdAuctionParamSource): Result<AdAuctionParams> {
         return auctionParamsScope {
@@ -55,17 +53,6 @@ internal class BigoAdsInterstitialImpl :
                 },
                 bidPrice = pricefloor,
             )
-        }
-    }
-
-    override suspend fun getToken(context: Context): String? = BigoAdSdk.getBidderToken()
-
-    override fun show(activity: Activity) {
-        val interstitialAd = interstitialAd
-        if (interstitialAd == null) {
-            emitEvent(AdEvent.ShowFailed(BidonError.AdNotReady))
-        } else {
-            interstitialAd.show()
         }
     }
 
@@ -92,23 +79,34 @@ internal class BigoAdsInterstitialImpl :
             .loadAd(builder.build())
     }
 
+    override fun show(activity: Activity) {
+        val interstitialAd = interstitialAd
+        if (interstitialAd == null) {
+            emitEvent(AdEvent.ShowFailed(BidonError.AdNotReady))
+        } else {
+            interstitialAd.show()
+        }
+    }
+
+    override fun destroy() {
+        interstitialAd?.destroy()
+        interstitialAd = null
+    }
+
     private fun fill(
         interstitialAd: InterstitialAd,
         adParams: BigoFullscreenAuctionParams
     ) {
-        val ad = getAd(this)
-        if (ad == null) {
-            emitEvent(AdEvent.ShowFailed(BidonError.AdNotReady))
-        } else {
-            interstitialAd.setAdInteractionListener(object : AdInteractionListener {
-                override fun onAdError(error: AdError) {
-                    val cause = error.asBidonError()
-                    logError(TAG, "onAdError: $this", cause)
-                    emitEvent(AdEvent.ShowFailed(cause))
-                }
+        interstitialAd.setAdInteractionListener(object : AdInteractionListener {
+            override fun onAdError(error: AdError) {
+                val cause = error.asBidonError()
+                logError(TAG, "onAdError: $this", cause)
+                emitEvent(AdEvent.ShowFailed(cause))
+            }
 
-                override fun onAdImpression() {
-                    logInfo(TAG, "onAdImpression: $this")
+            override fun onAdImpression() {
+                logInfo(TAG, "onAdImpression: $this")
+                getAd()?.let { ad ->
                     emitEvent(
                         AdEvent.PaidRevenue(
                             ad = ad,
@@ -120,22 +118,31 @@ internal class BigoAdsInterstitialImpl :
                         )
                     )
                 }
+            }
 
-                override fun onAdClicked() {
-                    logInfo(TAG, "onAdClicked: $this")
+            override fun onAdClicked() {
+                logInfo(TAG, "onAdClicked: $this")
+                getAd()?.let { ad ->
                     emitEvent(AdEvent.Clicked(ad))
                 }
+            }
 
-                override fun onAdOpened() {
-                    logInfo(TAG, "onAdOpened: $this")
+            override fun onAdOpened() {
+                logInfo(TAG, "onAdOpened: $this")
+                getAd()?.let { ad ->
                     emitEvent(AdEvent.Shown(ad))
                 }
+            }
 
-                override fun onAdClosed() {
-                    logInfo(TAG, "onAdClosed: $this")
+            override fun onAdClosed() {
+                logInfo(TAG, "onAdClosed: $this")
+                getAd()?.let { ad ->
                     emitEvent(AdEvent.Closed(ad))
                 }
-            })
+                this@BigoAdsInterstitialImpl.interstitialAd = null
+            }
+        })
+        getAd()?.let { ad ->
             emitEvent(AdEvent.Fill(ad))
         }
     }