+++ adapter/dtexchange/src/main/java/org/bidon/dtexchange/impl/DTExchangeInterstitial.kt
@@ -19,10 +19,8 @@ import org.bidon.sdk.adapter.AdSource
 import org.bidon.sdk.adapter.Mode
 import org.bidon.sdk.adapter.impl.AdEventFlow
 import org.bidon.sdk.adapter.impl.AdEventFlowImpl
-import org.bidon.sdk.ads.Ad
 import org.bidon.sdk.auction.models.LineItem
 import org.bidon.sdk.config.BidonError
-import org.bidon.sdk.logs.analytic.AdValue
 import org.bidon.sdk.logs.logging.impl.logInfo
 import org.bidon.sdk.stats.StatisticsCollector
 import org.bidon.sdk.stats.impl.StatisticsCollectorImpl
@@ -38,6 +36,7 @@ internal class DTExchangeInterstitial :
 
     private var lineItem: LineItem? = null
     private var inneractiveAdSpot: InneractiveAdSpot? = null
+    private var demandSource: String? = null
 
     override val isAdReadyToShow: Boolean
         get() = inneractiveAdSpot?.isReady == true
@@ -63,7 +62,9 @@ internal class DTExchangeInterstitial :
             ) {
                 logInfo(TAG, "onAdImpression: $adSpot")
                 val adValue = impressionData?.asAdValue() ?: return
-                val ad = adSpot?.asAd(impressionData.demandSource) ?: return
+                demandSource = impressionData.demandSource
+                setDsp(demandSource)
+                val ad = getAd() ?: return
                 emitEvent(AdEvent.PaidRevenue(ad, adValue))
                 emitEvent(AdEvent.Shown(ad))
             }
@@ -72,7 +73,7 @@ internal class DTExchangeInterstitial :
 
             override fun onAdClicked(adSpot: InneractiveAdSpot?) {
                 logInfo(TAG, "onAdClicked: $adSpot")
-                adSpot?.asAd()?.let {
+                getAd()?.let {
                     emitEvent(AdEvent.Clicked(ad = it))
                 }
             }
@@ -90,9 +91,10 @@ internal class DTExchangeInterstitial :
 
             override fun onAdDismissed(adSpot: InneractiveAdSpot?) {
                 logInfo(TAG, "onAdDismissed: $adSpot")
-                adSpot?.asAd()?.let {
+                getAd()?.let {
                     emitEvent(AdEvent.Closed(ad = it))
                 }
+                this@DTExchangeInterstitial.inneractiveAdSpot = null
             }
         }
         spot.addUnitController(controller)
@@ -102,8 +104,9 @@ internal class DTExchangeInterstitial :
                 override fun onInneractiveSuccessfulAdRequest(inneractiveAdSpot: InneractiveAdSpot?) {
                     logInfo(TAG, "onInneractiveSuccessfulAdRequest: $inneractiveAdSpot")
                     this@DTExchangeInterstitial.inneractiveAdSpot = inneractiveAdSpot
-                    inneractiveAdSpot?.let {
-                        emitEvent(AdEvent.Fill(it.asAd()))
+                    setDsp(demandSource ?: inneractiveAdSpot?.mediationNameString)
+                    getAd()?.let {
+                        emitEvent(AdEvent.Fill(it))
                     }
                 }
 
@@ -133,19 +136,6 @@ internal class DTExchangeInterstitial :
         inneractiveAdSpot?.destroy()
         inneractiveAdSpot = null
     }
-
-    private fun InneractiveAdSpot.asAd(demandSource: String? = null) = Ad(
-        ecpm = lineItem?.pricefloor ?: 0.0,
-        auctionId = auctionId,
-        adUnitId = lineItem?.adUnitId,
-        networkName = demandId.demandId,
-        currencyCode = AdValue.USD,
-        demandAd = demandAd,
-        dsp = demandSource ?: this.mediationNameString,
-        roundId = roundId,
-        demandAdObject = this,
-        bidType = bidType,
-    )
 }
 
 private const val TAG = "DTExchangeInterstitial"