+++ adapter/meta/src/main/java/org/bidon/meta/impl/MetaRewardedAdImpl.kt
@@ -15,6 +15,7 @@ import org.bidon.sdk.adapter.AdSource
 import org.bidon.sdk.adapter.Mode
 import org.bidon.sdk.adapter.impl.AdEventFlow
 import org.bidon.sdk.adapter.impl.AdEventFlowImpl
+import org.bidon.sdk.auction.AdTypeParam
 import org.bidon.sdk.config.BidonError
 import org.bidon.sdk.logs.analytic.AdValue
 import org.bidon.sdk.logs.analytic.Precision
@@ -37,7 +38,7 @@ class MetaRewardedAdImpl :
     override val isAdReadyToShow: Boolean
         get() = rewardedVideoAd?.isAdLoaded ?: false
 
-    override suspend fun getToken(context: Context): String? {
+    override suspend fun getToken(context: Context, adTypeParam: AdTypeParam): String? {
         return BidderTokenProvider.getBidderToken(context)
     }
 
@@ -73,7 +74,7 @@ class MetaRewardedAdImpl :
 
                     override fun onAdLoaded(ad: Ad?) {
                         logInfo(TAG, "onAdLoaded $ad: $rewardedVideoAd, $this")
-                        val bidonAd = getAd(this@MetaRewardedAdImpl)
+                        val bidonAd = getAd()
                         if (rewardedVideoAd != null && bidonAd != null) {
                             emitEvent(AdEvent.Fill(bidonAd))
                         } else {
@@ -83,13 +84,13 @@ class MetaRewardedAdImpl :
 
                     override fun onAdClicked(ad: Ad?) {
                         logInfo(TAG, "onAdClicked: $this")
-                        val bidonAd = getAd(this@MetaRewardedAdImpl) ?: return
+                        val bidonAd = getAd() ?: return
                         emitEvent(AdEvent.Clicked(bidonAd))
                     }
 
                     override fun onLoggingImpression(ad: Ad?) {
                         logInfo(TAG, "onAdImpression: $this")
-                        val bidonAd = getAd(this@MetaRewardedAdImpl) ?: return
+                        val bidonAd = getAd() ?: return
                         emitEvent(
                             AdEvent.PaidRevenue(
                                 ad = bidonAd,
@@ -104,15 +105,16 @@ class MetaRewardedAdImpl :
 
                     override fun onRewardedVideoCompleted() {
                         logInfo(TAG, "onRewardedVideoCompleted")
-                        val bidonAd = getAd(this@MetaRewardedAdImpl) ?: return
+                        val bidonAd = getAd() ?: return
                         emitEvent(AdEvent.Shown(bidonAd))
                         emitEvent(AdEvent.OnReward(bidonAd, null))
                     }
 
                     override fun onRewardedVideoClosed() {
                         logInfo(TAG, "onRewardedVideoClosed")
-                        val bidonAd = getAd(this@MetaRewardedAdImpl) ?: return
+                        val bidonAd = getAd() ?: return
                         emitEvent(AdEvent.Closed(bidonAd))
+                        this@MetaRewardedAdImpl.rewardedVideoAd = null
                     }
                 })
                 .withBid(adParams.payload)