+++ adapter/mobilefuse/src/main/java/org/bidon/mobilefuse/impl/MobileFuseInterstitialImpl.kt
@@ -1,0 +1,148 @@
+package org.bidon.mobilefuse.impl
+
+import android.app.Activity
+import android.content.Context
+import com.mobilefuse.sdk.AdError
+import com.mobilefuse.sdk.MobileFuseInterstitialAd
+import kotlinx.coroutines.flow.MutableSharedFlow
+import org.bidon.mobilefuse.ext.GetMobileFuseTokenUseCase
+import org.bidon.sdk.adapter.AdAuctionParamSource
+import org.bidon.sdk.adapter.AdAuctionParams
+import org.bidon.sdk.adapter.AdEvent
+import org.bidon.sdk.adapter.AdSource
+import org.bidon.sdk.adapter.Mode
+import org.bidon.sdk.adapter.impl.AdEventFlow
+import org.bidon.sdk.adapter.impl.AdEventFlowImpl
+import org.bidon.sdk.auction.AdTypeParam
+import org.bidon.sdk.config.BidonError
+import org.bidon.sdk.logs.analytic.AdValue
+import org.bidon.sdk.logs.analytic.AdValue.Companion.USD
+import org.bidon.sdk.logs.analytic.Precision
+import org.bidon.sdk.logs.logging.impl.logError
+import org.bidon.sdk.logs.logging.impl.logInfo
+import org.bidon.sdk.stats.StatisticsCollector
+import org.bidon.sdk.stats.impl.StatisticsCollectorImpl
+import java.util.concurrent.atomic.AtomicBoolean
+
+class MobileFuseInterstitialImpl(private val isTestMode: Boolean) :
+    AdSource.Interstitial<MobileFuseFullscreenAuctionParams>,
+    Mode.Bidding,
+    AdEventFlow by AdEventFlowImpl(),
+    StatisticsCollector by StatisticsCollectorImpl() {
+
+    private var interstitialAd: MobileFuseInterstitialAd? = null
+
+    /**
+     * This flag is used to prevent [AdError]-callback from being exposed twice.
+     */
+    private var isLoaded = AtomicBoolean(false)
+
+    override val adEvent = MutableSharedFlow<AdEvent>(extraBufferCapacity = Int.MAX_VALUE, replay = 1)
+    override val isAdReadyToShow: Boolean get() = interstitialAd?.isLoaded == true
+
+    override suspend fun getToken(context: Context, adTypeParam: AdTypeParam): String? {
+        return GetMobileFuseTokenUseCase(context, isTestMode)
+    }
+
+    override fun getAuctionParam(auctionParamsScope: AdAuctionParamSource): Result<AdAuctionParams> {
+        return ObtainAuctionParamUseCase().getFullscreenParam(auctionParamsScope)
+    }
+
+    override fun load(adParams: MobileFuseFullscreenAuctionParams) {
+        logInfo(Tag, "Starting with $adParams: $this")
+        // placementId should be configured in the mediation platform UI and passed back to this method:
+        val interstitialAd = MobileFuseInterstitialAd(adParams.activity, adParams.placementId).also {
+            interstitialAd = it
+        }
+        interstitialAd.setListener(object : MobileFuseInterstitialAd.Listener {
+            override fun onAdLoaded() {
+                if (!isLoaded.getAndSet(true)) {
+                    logInfo(Tag, "onAdLoaded")
+                    getAd()?.let { adEvent.tryEmit(AdEvent.Fill(it)) }
+                }
+            }
+
+            override fun onAdNotFilled() {
+                val cause = BidonError.NoFill(demandId)
+                logError(Tag, "onAdNotFilled", cause)
+                adEvent.tryEmit(AdEvent.LoadFailed(cause))
+            }
+
+            override fun onAdRendered() {
+                logInfo(Tag, "onAdRendered")
+                getAd()?.let {
+                    adEvent.tryEmit(AdEvent.Shown(it))
+                    adEvent.tryEmit(
+                        AdEvent.PaidRevenue(
+                            ad = it,
+                            adValue = interstitialAd.winningBidInfo.let { bidInfo ->
+                                AdValue(
+                                    adRevenue = bidInfo?.cpmPrice?.toDouble() ?: 0.0,
+                                    currency = bidInfo?.currency ?: USD,
+                                    precision = Precision.Precise
+                                )
+                            }
+                        )
+                    )
+                }
+            }
+
+            override fun onAdClicked() {
+                logInfo(Tag, "onAdClicked")
+                getAd()?.let { adEvent.tryEmit(AdEvent.Clicked(it)) }
+            }
+
+            override fun onAdExpired() {
+                logInfo(Tag, "onAdExpired")
+                adEvent.tryEmit(AdEvent.LoadFailed(BidonError.Expired(demandId)))
+            }
+
+            override fun onAdError(adError: AdError?) {
+                logError(Tag, "onAdError $adError", Throwable(adError?.errorMessage))
+                when (adError) {
+                    AdError.AD_ALREADY_RENDERED -> {
+                        adEvent.tryEmit(AdEvent.ShowFailed(BidonError.AdNotReady))
+                    }
+
+                    AdError.AD_ALREADY_LOADED,
+                    AdError.AD_RUNTIME_ERROR -> {
+                        // do nothing
+                    }
+
+                    AdError.AD_LOAD_ERROR -> {
+                        if (!isLoaded.getAndSet(true)) {
+                            getAd()?.let { adEvent.tryEmit(AdEvent.LoadFailed(BidonError.NoFill(demandId))) }
+                        }
+                    }
+
+                    else -> {
+                        // do nothing
+                    }
+                }
+            }
+
+            override fun onAdClosed() {
+                logInfo(Tag, "onAdClosed: $this")
+                getAd()?.let { adEvent.tryEmit(AdEvent.Closed(it)) }
+                this@MobileFuseInterstitialImpl.interstitialAd = null
+            }
+        })
+        interstitialAd.loadAdFromBiddingToken(adParams.signalData)
+    }
+
+    override fun show(activity: Activity) {
+        logInfo(Tag, "Starting show: $this")
+        if (interstitialAd?.isLoaded == true) {
+            interstitialAd?.showAd()
+        } else {
+            adEvent.tryEmit(AdEvent.ShowFailed(BidonError.AdNotReady))
+        }
+    }
+
+    override fun destroy() {
+        logInfo(Tag, "destroy $this")
+        interstitialAd = null
+    }
+}
+
+private const val Tag = "MobileFuseInterstitialImpl"